requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
openai>=1.3.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
chromadb>=0.4.18
python-dotenv>=1.0.0
python-multipart>=0.0.6







